Extract GitHub artifact helpers
