+++ tools/flakereport/github.go
@@ -11,6 +11,8 @@ import (
 	"path/filepath"
 	"strings"
 	"time"
+
+	"go.temporal.io/server/tools/common/github"
 )
 
 // fetchWorkflowRuns retrieves all completed workflow runs within a date range.
@@ -72,30 +74,15 @@ func fetchWorkflowRuns(ctx context.Context, repo string, workflowID int64, branc
 }
 
 // fetchRunArtifacts retrieves all artifacts for a specific workflow run
-func fetchRunArtifacts(ctx context.Context, repo string, runID int64) ([]WorkflowArtifact, error) {
-	ctxTimeout, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-
-	cmd := exec.CommandContext(ctxTimeout, "gh", "api",
-		fmt.Sprintf("/repos/%s/actions/runs/%d/artifacts?per_page=100", repo, runID),
-	)
-
-	output, err := cmd.Output()
+func fetchRunArtifacts(ctx context.Context, repo string, runID int64) ([]github.Artifact, error) {
+	artifacts, err := github.ListRunArtifacts(ctx, repo, runID)
 	if err != nil {
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			return nil, fmt.Errorf("gh api failed for run %d: %w\nstderr: %s", runID, err, string(exitErr.Stderr))
-		}
-		return nil, fmt.Errorf("failed to execute gh command for run %d: %w", runID, err)
-	}
-
-	var response ArtifactsResponse
-	if err := json.Unmarshal(output, &response); err != nil {
-		return nil, fmt.Errorf("failed to parse artifacts response for run %d: %w", runID, err)
+		return nil, err
 	}
 
 	// Filter for JUnit/test artifacts
-	var testArtifacts []WorkflowArtifact
-	for _, artifact := range response.Artifacts {
+	var testArtifacts []github.Artifact
+	for _, artifact := range artifacts {
 		if artifact.Expired {
 			continue
 		}
@@ -106,32 +93,6 @@ func fetchRunArtifacts(ctx context.Context, repo string, runID int64) ([]Workflo
 	}
 
 	return testArtifacts, nil
-}
-
-// downloadArtifact downloads a single artifact zip file
-func downloadArtifact(ctx context.Context, repo string, artifactID int64, outputDir string) (string, error) {
-	ctxTimeout, cancel := context.WithTimeout(ctx, 60*time.Second)
-	defer cancel()
-
-	zipPath := filepath.Join(outputDir, fmt.Sprintf("artifact-%d.zip", artifactID))
-
-	cmd := exec.CommandContext(ctxTimeout, "gh", "api",
-		fmt.Sprintf("/repos/%s/actions/artifacts/%d/zip", repo, artifactID),
-	)
-
-	output, err := cmd.Output()
-	if err != nil {
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			return "", fmt.Errorf("failed to download artifact %d: %w\nstderr: %s", artifactID, err, string(exitErr.Stderr))
-		}
-		return "", fmt.Errorf("failed to download artifact %d: %w", artifactID, err)
-	}
-
-	if err := os.WriteFile(zipPath, output, 0644); err != nil {
-		return "", fmt.Errorf("failed to write artifact zip %d: %w", artifactID, err)
-	}
-
-	return zipPath, nil
 }
 
 // extractArtifactZip extracts zip file and returns paths to JUnit XML files
